Refactored bootstrapping into core, service actions, parsing and actual bootstrapping
