+++ src/Game/CMakeLists.txt
@@ -1,0 +1,10 @@
+# Add game-specific code
+
+# - Parsing utilities
+target_sources(game PRIVATE Parsing.cpp)
+
+# - Service actions
+target_sources(game PRIVATE ServiceActions.cpp)
+
+# - Bootstrapping
+target_sources(game PRIVATE main.cpp)
